rectify ray_sklearn + remove predict_proba for sklearn, will keep only for keras
